requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
coverage==7.9.1
Django==5.2.3
djangorestframework==3.16.0
idna==3.10
PyJWT==2.10.1
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
